continue removing manager
